+++ src/ffi2/mod.rs
@@ -16,7 +16,6 @@ pub use self::cobject::*;
 pub use self::code::*;
 pub use self::compile::*;
 pub use self::complexobject::*;
-pub use self::datetime::*;
 pub use self::descrobject::*;
 pub use self::dictobject::*;
 pub use self::enumobject::*;
@@ -67,7 +66,6 @@ mod cellobject;
 mod classobject;
 mod cobject;
 mod complexobject;
-mod datetime;
 mod descrobject;
 mod dictobject;
 mod enumobject;
